Move types to files
