+++ src/GraphQL.Tests/Utilities/FederatedSchemaBuilderTestBase.cs
@@ -1,0 +1,12 @@
+using GraphQL.Utilities.Federation;
+
+namespace GraphQL.Tests.Utilities
+{
+    public class FederatedSchemaBuilderTestBase : SchemaBuilderTestBase
+    {
+        public FederatedSchemaBuilderTestBase()
+        {
+            Builder = new FederatedSchemaBuilder();
+        }
+    }
+}
