+++ src/GraphQL/Utilities/Federation/IFederatedResolver.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace GraphQL.Utilities.Federation
+{
+    public interface IFederatedResolver
+    {
+        Task<object> Resolve(FederatedResolveContext context);
+    }
+}
